+++ transports/deflate/tests/test.rs
@@ -44,7 +44,7 @@ fn lot_of_data() {
 }
 
 async fn run(message1: Vec<u8>) {
-    let transport = TcpConfig::new().and_then(|conn, endpoint| {
+    let mut transport = TcpConfig::new().and_then(|conn, endpoint| {
         upgrade::apply(
             conn,
             DeflateConfig::default(),