+++ protocols/ping/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.36.0 [unreleased]
+
+- Update to `libp2p-core` `v0.33.0`.
+
+- Update to `libp2p-swarm` `v0.36.0`.
+
 # 0.35.0
 
 - Update to `libp2p-swarm` `v0.35.0`.